evals/mirascope_skill/pyproject.toml
[project]
name = "mirascope-skill-eval"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "mirascope[all]",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]
